Load TINC input data.

The function loads input mutations and optional copy number data for TINC. Input formats are reported at the package website.

After loading data, mutation with VAF outside a range (default [0; 0.7]) are removed from analysis. Similarly, mutations are down-sampled if in excess of some threshold (default `N=20000`).

load_TINC_input(x, cna, VAF_range_tumour = c(0, 0.7), N = 20000)

Arguments

x
A dataframe or tibble with input mutation data, reporting `chr`, `from`, `to`, `ref` and `alt`, plus `n_ref_count` and `n_alt_count`, and `t_ref_count` and `t_tot_count`.

cna
Copy Number data in the format of package CNAqc, providing `chr`, `from`, `to`, `Major` and `minor`.

VAF_range_tumour
VAF range used to filter mutations from the tumour sample.

N
If there are more than `N` mutations in VAF range `VAF_range_tumour`, a random subset of size `N` is retained.

Value

A tibble with the loaded data.

Examples

# Generating a random TIN input
rt = random_TIN()

#> ✓ Generated TINC dataset (n = 997 mutations), TIN (0.05) and TIT (1), normal and tumour coverage 30x and 120x.

#> Warning: Removed 2 rows containing missing values (geom_bar).

#> Warning: Removed 2 rows containing missing values (geom_bar).

load_TINC_input(x = rt$data, cna = rt$cna)

#> 

#> ── Loading TINC input data ─────────────────────────────────────────────────────

#> ✓ Input data contains n = 997 mutations, selecting operation mode.

#> ! Found CNA data, retaining only mutations that map to segments with predominant karyotype ...

#> 
#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ! Missing segments length from CNA calls, adding it to CNA calls.

#> ℹ Input n = 997 mutations for 997 CNA segments (997 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> 

#> ✓ Mapped n = 997 mutations to clonal segments (100% of input)

#> 

#> 

#> ── Genome coverage by karyotype, in basepairs. ──

#> 

#> # A tibble: 1 x 4
#>   minor Major     n karyotype
#>   <dbl> <dbl> <dbl> <chr>    
#> 1     1     1  2991 1:1      

#> ✓ n = 997 mutations mapped to CNA segments with karyotype 1:1 (largest available in basepairs).

#> ✓ Mutation with VAF within 0 and 0.7 ~ n = 995.

#> $mutations
#> # A tibble: 997 x 12
#>    chr        from        to ref   alt   n_ref_count n_alt_count t_ref_count
#>    <chr>     <int>     <dbl> <chr> <chr>       <dbl>       <dbl>       <dbl>
#>  1 chr15  29675019  29675020 T     C              25           0          53
#>  2 chr7  127986905 127986906 G     G              21           0          67
#>  3 chr10  84255144  84255145 A     A              24           0          53
#>  4 chr19  38367178  38367179 T     A              25           0          62
#>  5 chr10  26886489  26886490 C     C              28           0          56
#>  6 chr12  10757148  10757149 C     C              33           0          42
#>  7 chr6   36997904  36997905 G     T              26           0          62
#>  8 chr8   56370544  56370545 G     C              37           0          70
#>  9 chr20  34992347  34992348 A     T              31           0          53
#> 10 chr5  116203285 116203286 G     G              32           0          57
#> # … with 987 more rows, and 4 more variables: t_alt_count <dbl>,
#> #   karyotype <chr>, id <chr>, OK_tumour <lgl>
#> 
#> $cna_map

#> ── [ CNAqc ]  n = 997 mutations in 997 segments (997 clonal + 0 subclonal). Geno

#> 
#>  1:1  [n = 997, L =   0 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■
#> 

#> ℹ Sample Purity: 80% ~ Ploidy: 2.

#> 
#> $what_we_used
#> [1] "1:1"
#>