Plot a TINC analysis.

This function is a wrapper to call plot_simple_report.

# S3 method for tin_obj
plot(x, ...)

Arguments

x
A TINC analysis computed with autofit.

...
Extra S3 parameters.

Value

A TINC analysis plot (a `ggplot` figure with multiple panels).

See also

Examples

rt = random_TIN()

#> ✓ Generated TINC dataset (n = 991 mutations), TIN (0.05) and TIT (1), normal and tumour coverage 30x and 120x.

#> Warning: Removed 2 rows containing missing values (geom_bar).

#> Warning: Removed 2 rows containing missing values (geom_bar).

plot(autofit(input = rt$data, cna = rt$cna, FAST = TRUE))

#>  [ TINC ] 
#> 

#> 

#> ── Loading TINC input data ─────────────────────────────────────────────────────

#> ✓ Input data contains n = 991 mutations, selecting operation mode.

#> ! Found CNA data, retaining only mutations that map to segments with predominant karyotype ...

#> 
#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ! Missing segments length from CNA calls, adding it to CNA calls.

#> ℹ Input n = 991 mutations for 991 CNA segments (991 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> 

#> ✓ Mapped n = 991 mutations to clonal segments (100% of input)

#> 

#> 

#> ── Genome coverage by karyotype, in basepairs. ──

#> 

#> # A tibble: 1 x 4
#>   minor Major     n karyotype
#>   <dbl> <dbl> <dbl> <chr>    
#> 1     1     1  2973 1:1      

#> ✓ n = 991 mutations mapped to CNA segments with karyotype 1:1 (largest available in basepairs).

#> ✓ Mutation with VAF within 0 and 0.7 ~ n = 989.

#> 

#> ── Analysing tumour sample with MOBSTER ────────────────────────────────────────

#> 
#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 989.

#> ❯ n = 989. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output
#> clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 
#> 
#> 

#> ℹ MOBSTER fits completed in 9s.

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 989 with k = 1 Beta(s) and a tail ──────────

#> ● Clusters: π = 62% [C1] and 38% [Tail], with π > 0.

#> ● Tail [n = 360, 38%] with alpha = 1.2.

#> ● Beta C1 [n = 629, 62%] with mean = 0.49.

#> ℹ Score(s): NLL = -975.82; ICL = -1843.73 (-1910.26), H = 66.53 (0). Fit
#> converged by MM in 15 steps.

#> 

#> ℹ With CNA, TINC will estimating tumour purity adjusting by copy number and mutation multiplicity.

#> ℹ Mutant allele copies 1 for karyotype 1:1

#> Warning: You did not pass enough input colours, adding a gray colour
#> Available: C1, Tail
#> Missing: NA

#> 

#> ✓ MOBSTER found n = 623 clonal mutations from cluster C1

#> 

#> ── Analysing normal sample with BMix ───────────────────────────────────────────

#> 

#> 

#> ── BMix fit ────────────────────────────────────────────────────────────────────

#> 

#> ℹ Binomials k_B = 1 and 2, Beta-Binomials k_BB = 0; 4 fits to run.

#> 

#> ℹ Bmix best fit completed in 0 mins

#> 

#> ── [ BMix ] My BMix model n = 623 with k = 2 component(s) (2 + 0) ──────────────

#> ● Clusters: π = 97% [Bin 1] and 3% [Bin 2], with π > 0.

#> ● Binomial Bin 1 with mean = 0.00152257480686904.

#> ● Binomial Bin 2 with mean = 0.0126173488233142.

#> ℹ Score (model selection): ICL = 391.77.

#> Scale for 'x' is already present. Adding another scale for 'x', which will
#> replace the existing scale.

#> Scale for 'fill' is already present. Adding another scale for 'fill', which
#> will replace the existing scale.

#> Warning: Removed 2 rows containing missing values (geom_bar).

#> Warning: Removed 551 rows containing missing values (geom_raster).

#> ✓ Binomial peaks 0.00152257480686904 and 0.0126173488233142 with proportions 0.966108255489676 and 0.0338917445103242. Clonal score 0.00189859605323418 with TINN 0.00379719210646837

#> 

#> ── Analysing tumour and normal samples with VIBER ──────────────────────────────

#> 
#>  [ VIBER - variational fit ] 
#> 

#> ℹ Input n = 989, with k < 5. Dirichlet concentration α = 1e-06.

#> ℹ Beta (a_0, b_0) = (1, 1); q_i = prior. Optimise: ε = 1e-06 or 1000 steps, r = 3 starts.

#> [easypar] 2021-04-22 01:10:25 - Overriding parallel execution setup [TRUE] with global option : FALSE

#> Warning: ELBO decreased by 4.37917115399614at step #6

#> 

#> ✓ VIBER fit completed in 0.01 mins (status: converged)

#> 

#> ── [ VIBER ] My VIBER model n = 989 (w = 2 dimensions). Fit with k = 5 clusters.

#> ● Clusters: π = 65% [C2], 30% [C1], and 5% [C5], with π > 0.

#> ● Binomials: θ = <0, 0.49> [C2], <0, 0.08> [C1], and <0, 0.23> [C5].

#> ℹ Score(s): ELBO = -132294.334. Fit converged in 38 steps, ε = 1e-06.

#> 

#> ✓ Reduced to k = 3 (from 5) selecting VIBER cluster(s) with π > 0.02, and Binomial p > 0 in w > 0 dimension(s).

#> Error in geom_label(data = vtin_g, aes(y = lab.ypos, label = paste0(round(value *     100, 1), "%")), color = "black", fill = "white"): object 'vtin_g' not found