Print a TINC object to screen

Print a TINC object to screen

# S3 method for tin_obj
print(x, ...)

Arguments

x
A TINC analysis computed with autofit.

...
Extra S3 parameters

Value

Nothing.

Examples

# Automatic call
rt = random_TIN()
#> ✔ Generated TINC dataset (n = 993 mutations), TIN (0.05) and TIT (1), normal and tumour coverage 30x and 120x.
#> Warning: Removed 2 rows containing missing values (geom_bar).
#> Warning: Removed 2 rows containing missing values (geom_bar).
autofit(input = rt$data, cna = rt$cna, FAST = TRUE)
#>  [ TINC ] 
#> 
#> 
#> ── Loading TINC input data ─────────────────────────────────────────────────────
#> ✔ Input data contains n = 993 mutations, selecting operation mode.
#> ! Found CNA data, retaining only mutations that map to segments with predominant karyotype ...
#> 
#>  [ CNAqc - CNA Quality Check ] 
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.
#> ! Missing segments length from CNA calls, adding it to CNA calls.
#> ℹ Input n = 993 mutations for 993 CNA segments (993 clonal, 0 subclonal)
#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten
#> 
#> ✔ Mapped n = 993 mutations to clonal segments (100% of input)
#> 
#> 
#> ── Genome coverage by karyotype, in basepairs. ──
#> 
#> # A tibble: 1 × 4
#>   minor Major     n karyotype
#>   <dbl> <dbl> <dbl> <chr>    
#> 1     1     1  2979 1:1      
#> ✔ n = 993 mutations mapped to CNA segments with karyotype 1:1 (largest available in basepairs).
#> ✔ Mutation with VAF within 0 and 0.7 ~ n = 989.
#> 
#> ── Analysing tumour sample with MOBSTER ────────────────────────────────────────
#> 
#>  [ MOBSTER fit ] 
#> 
#> ✔ Loaded input data, n = 989.
#> ❯ n = 989. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output
#> clusters with π > 0.02 and n > 10.
#> ! mobster automatic setup FAST for the analysis.
#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.
#> 
#> 
#> 
#> ℹ MOBSTER fits completed in 14.2s.
#> 
#> ── [ MOBSTER ] My MOBSTER model n = 989 with k = 1 Beta(s) and a tail ──────────
#> ● Clusters: π = 70% [C1] and 30% [Tail], with π > 0.
#> ● Tail [n = 284, 30%] with alpha = 1.2.
#> ● Beta C1 [n = 705, 70%] with mean = 0.5.
#> ℹ Score(s): NLL = -1065.96; ICL = -2036.5 (-2090.53), H = 54.04 (0). Fit
#> converged by MM in 12 steps.
#> 
#> ℹ With CNA, TINC will estimating tumour purity adjusting by copy number and mutation multiplicity.
#> ℹ Mutant allele copies 1 for karyotype 1:1
#> Warning: You did not pass enough input colours, adding a gray colour
#> Available: C1, Tail
#> Missing: NA
#> Warning: `guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
#> 
#> ✔ MOBSTER found n = 699 clonal mutations from cluster C1
#> 
#> ── Analysing normal sample with BMix ───────────────────────────────────────────
#> 
#> 
#> ── BMix fit ────────────────────────────────────────────────────────────────────
#> 
#> ℹ Binomials k_B = 1 and 2, Beta-Binomials k_BB = 0; 4 fits to run.
#> 
#> ℹ Bmix best fit completed in 0.01 mins
#> 
#> ── [ BMix ] My BMix model n = 699 with k = 1 component(s) (1 + 0) ──────────────
#> • Clusters: π = 100% [Bin 1], with π > 0.
#> • Binomial Bin 1 with mean = 0.00232657518636342.
#> ℹ Score (model selection): ICL = 811.26.
#> Scale for 'x' is already present. Adding another scale for 'x', which will
#> replace the existing scale.
#> Scale for 'fill' is already present. Adding another scale for 'fill', which
#> will replace the existing scale.
#> Warning: Removed 2 rows containing missing values (geom_bar).
#> Warning: Removed 551 rows containing missing values (geom_raster).
#> ✔ Binomial peaks 0.00232657518636342 with proportions 1. Clonal score 0.00232657518636342 with TINN 0.00465315037272684
#> 
#> ── Analysing tumour and normal samples with VIBER ──────────────────────────────
#> 
#>  [ VIBER - variational fit ] 
#> 
#> ℹ Input n = 989, with k < 5. Dirichlet concentration α = 1e-06.
#> ℹ Beta (a_0, b_0) = (1, 1); q_i = prior. Optimise: ε = 1e-06 or 1000 steps, r = 3 starts.
#> [easypar] 2021-11-22 13:51:53 - Overriding parallel execution setup [TRUE] with global option : FALSE
#> 
#> ✔ VIBER fit completed in 0.02 mins (status: converged)
#> 
#> ── [ VIBER ] My VIBER model n = 989 (w = 2 dimensions). Fit with k = 5 clusters.
#> • Clusters: π = 72% [C4], 24% [C5], and 4% [C3], with π > 0.
#> • Binomials: θ = <0, 0.5> [C4], <0, 0.08> [C5], and <0, 0.21> [C3].
#> ℹ Score(s): ELBO = -138229.443. Fit converged in 25 steps, ε = 1e-06.
#> 
#> ✔ Reduced to k = 3 (from 5) selecting VIBER cluster(s) with π > 0.02, and Binomial p > 0 in w > 0 dimension(s).
#> 
#> ── TINC profiler for bulk WGS ──────────────────────────────────────────────────
#> 
#> ℹ Copy Number data has been used for this analysis (karyotype 1:1)
#> 
#> ────────────────────────────────────────────────────────────────────────────────
#> ── [ CNAqc ]  n = 993 mutations in 993 segments (993 clonal + 0 subclonal). Geno
#> 
#>  1:1  [n = 993, L =   0 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■
#> 
#> ℹ Sample Purity: 80% ~ Ploidy: 2.
#> ────────────────────────────────────────────────────────────────────────────────
#> → Mutations data: n = 989 out of 993 within range (100%).
#> 
#>         TIT :  100% (RF 100)  ~ n = 699 clonal mutations, cluster C1 
#>         TIN :  0% (RF 0)  ~ n = 49 with VAF > 0 
#> 
#>    QC Tumour   High purity (>85%)
#>    QC Normal   No Contamination (<1%)