Runs a TINC analysis.

This function is a wrapper to run the main analysis of TINC.

The steps are as follows:

1) Input data is loaded from a `file`, or from a `dataframe`.

2) Clonal mutations are estimated for the tumour, together with the tumour purity (Tumour in Tumour).

3) From putative clonal mutations of the tumour, the Tumour in Normal contamination level is estimated.

An S3 object is returned that contains the results of the analysis.

autofit(
  input,
  cna,
  VAF_range_tumour = c(0, 0.7),
  cutoff_miscalled_clonal = 0.6,
  cutoff_lv_assignment = 0.75,
  N = 20000,
  FAST = FALSE
)

Arguments

input
A`dataframe` of the iput mutations. Must be in a certain format, see the vignette for more information.

cna
Copy Number data in the format of package CNAqc.

VAF_range_tumour
A range `[x, y]` so that only mutations with VAF in that range are actually used to determine the TIN/ TIT levels of the input.

cutoff_miscalled_clonal
An upper bound on the VAF of a cluster in the tumour data. Clusters above this value will be considered miscalled clonal clusters (e.g., due to LOH etc.).

cutoff_lv_assignment
Consider only latent variables with responsibilities above this cutoff.

N
If there are more than `N` mutations in VAF range `VAF_range_tumour`, a random subset of size `N` is retained.

FAST
If `TRUE`, it runs the analysis with reduced sampling power and accuracy. Use this to obtain a result for preliminary inspection of your data, and then run `autofit` with this parameter set to `FALSE`.

Value

An S3 object that contains the results of this analysis.

Examples

# Random
rt = random_TIN()
#> ✓ Generated TINC dataset (n = 994 mutations), TIN (0.05) and TIT (1), normal and tumour coverage 30x and 120x.
#> Warning: Removed 2 rows containing missing values (geom_bar).
#> Warning: Removed 2 rows containing missing values (geom_bar).

x = autofit(input = rt$data, cna = rt$cna, FAST = TRUE)
#>  [ TINC ] 
#> 
#> 
#> ── Loading TINC input data ─────────────────────────────────────────────────────
#> ✓ Input data contains n = 994 mutations, selecting operation mode.
#> ! Found CNA data, retaining only mutations that map to segments with predominant karyotype ...
#> 
#>  [ CNAqc - CNA Quality Check ] 
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.
#> ! Missing segments length from CNA calls, adding it to CNA calls.
#> ℹ Input n = 994 mutations for 994 CNA segments (994 clonal, 0 subclonal)
#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten
#> 
#> ✓ Mapped n = 994 mutations to clonal segments (100% of input)
#> 
#> 
#> ── Genome coverage by karyotype, in basepairs. ──
#> 
#> # A tibble: 1 × 4
#>   minor Major     n karyotype
#>   <dbl> <dbl> <dbl> <chr>    
#> 1     1     1  2982 1:1      
#> ✓ n = 994 mutations mapped to CNA segments with karyotype 1:1 (largest available in basepairs).
#> ✓ Mutation with VAF within 0 and 0.7 ~ n = 992.
#> 
#> ── Analysing tumour sample with MOBSTER ────────────────────────────────────────
#> 
#>  [ MOBSTER fit ] 
#> 
#> ✔ Loaded input data, n = 992.
#> ❯ n = 992. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output
#> clusters with π > 0.02 and n > 10.
#> ! mobster automatic setup FAST for the analysis.
#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.
#> 
#> 
#> 
#> ℹ MOBSTER fits completed in 3.3s.
#> 
#> ── [ MOBSTER ] My MOBSTER model n = 992 with k = 1 Beta(s) and a tail ──────────
#> ● Clusters: π = 65% [C1] and 35% [Tail], with π > 0.
#> ● Tail [n = 335, 35%] with alpha = 1.2.
#> ● Beta C1 [n = 657, 65%] with mean = 0.5.
#> ℹ Score(s): NLL = -1090.79; ICL = -2087.36 (-2140.18), H = 52.82 (0). Fit
#> converged by MM in 12 steps.
#> 
#> ℹ With CNA, TINC will estimating tumour purity adjusting by copy number and mutation multiplicity.
#> ℹ Mutant allele copies 1 for karyotype 1:1
#> Warning: You did not pass enough input colours, adding a gray colour
#> Available: C1, Tail
#> Missing: NA
#> Warning: `guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
#> 
#> ✓ MOBSTER found n = 654 clonal mutations from cluster C1
#> 
#> ── Analysing normal sample with BMix ───────────────────────────────────────────
#> 
#> 
#> ── BMix fit ────────────────────────────────────────────────────────────────────
#> 
#> ℹ Binomials k_B = 1 and 2, Beta-Binomials k_BB = 0; 4 fits to run.
#> 
#> ℹ Bmix best fit completed in 0 mins
#> 
#> ── [ BMix ] My BMix model n = 654 with k = 1 component(s) (1 + 0) ──────────────
#> • Clusters: π = 100% [Bin 1], with π > 0.
#> • Binomial Bin 1 with mean = 0.00210019465218728.
#> ℹ Score (model selection): ICL = 742.76.
#> Scale for 'x' is already present. Adding another scale for 'x', which will
#> replace the existing scale.
#> Scale for 'fill' is already present. Adding another scale for 'fill', which
#> will replace the existing scale.
#> Warning: Removed 2 rows containing missing values (geom_bar).
#> Warning: Removed 551 rows containing missing values (geom_raster).
#> ✓ Binomial peaks 0.00210019465218728 with proportions 1. Clonal score 0.00210019465218728 with TINN 0.00420038930437455
#> 
#> ── Analysing tumour and normal samples with VIBER ──────────────────────────────
#> 
#>  [ VIBER - variational fit ] 
#> 
#> ℹ Input n = 992, with k < 5. Dirichlet concentration α = 1e-06.
#> ℹ Beta (a_0, b_0) = (1, 1); q_i = prior. Optimise: ε = 1e-06 or 1000 steps, r = 3 starts.
#> [easypar] 2021-12-21 18:16:02 - Overriding parallel execution setup [TRUE] with global option : FALSE
#> 
#> ✓ VIBER fit completed in 0.01 mins (status: converged)
#> 
#> ── [ VIBER ] My VIBER model n = 992 (w = 2 dimensions). Fit with k = 5 clusters.
#> • Clusters: π = 68% [C3], 27% [C4], and 5% [C1], with π > 0.
#> • Binomials: θ = <0, 0.5> [C3], <0, 0.08> [C4], and <0, 0.22> [C1].
#> ℹ Score(s): ELBO = -135668.54. Fit converged in 40 steps, ε = 1e-06.
#> 
#> ✓ Reduced to k = 3 (from 5) selecting VIBER cluster(s) with π > 0.02, and Binomial p > 0 in w > 0 dimension(s).
#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
#> Using compatibility `.name_repair`.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.

print(x)
#> 
#> ── TINC profiler for bulk WGS ──────────────────────────────────────────────────
#> 
#> ℹ Copy Number data has been used for this analysis (karyotype 1:1)
#> 
#> ────────────────────────────────────────────────────────────────────────────────
#> ── [ CNAqc ]  n = 994 mutations in 994 segments (994 clonal + 0 subclonal). Geno
#> 
#>  1:1  [n = 994, L =   0 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■
#> 
#> ℹ Sample Purity: 80% ~ Ploidy: 2.
#> ────────────────────────────────────────────────────────────────────────────────
#> → Mutations data: n = 992 out of 994 within range (100%).
#> 
#>         TIT :  101% (RF 101)  ~ n = 654 clonal mutations, cluster C1 
#>         TIN :  0% (RF 0)  ~ n = 41 with VAF > 0 
#> 
#>    QC Tumour   High purity (>85%)
#>    QC Normal   No Contamination (<1%)

# Fit in the package
data('fit_example', package = 'TINC')

print(fit_example)
#> 
#> ── TINC profiler for bulk WGS ──────────────────────────────────────────────────
#> 
#> ℹ Copy Number data has been used for this analysis (karyotype 1:1)
#> 
#> ────────────────────────────────────────────────────────────────────────────────
#> ── [ CNAqc ]  n = 985 mutations in 985 segments (985 clonal + 0 subclonal). Geno
#> 
#>  1:1  [n = 985, L =   0 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■
#> 
#> ℹ Sample Purity: 80% ~ Ploidy: 2.
#> ────────────────────────────────────────────────────────────────────────────────
#> → Mutations data: n = 985 out of 985 within range (100%).
#> 
#>         TIT :  78% (RF 78)  ~ n = 714 clonal mutations, cluster C1 
#>         TIN :  7% (RF 7)  ~ n = 694 with VAF > 0 
#> 
#>    QC Tumour   Good purity (65-85%)
#>    QC Normal   Some contamination (3-7%)