Runs a TINC analysis.

This function is a wrapper to run the main analysis of TINC.

The steps are as follows:

1) Input data is loaded from a `file`, or from a `dataframe`.

2) Clonal mutations are estimated for the tumour, together with the tumour purity (Tumour in Tumour).

3) From putative clonal mutations of the tumour, the Tumour in Normal contamination level is estimated.

An S3 object is returned that contains the results of the analysis.

autofit(input, VAF_range_tumour = c(0, 0.7),
  cutoff_miscalled_clonal = 0.6, cutoff_lv_assignment = 0, N = 20000,
  FAST = FALSE)

Arguments

input
A `file` to load from disk, or a `dataframe `consistent with the required input. See the vignette for more information.

VAF_range_tumour
A range `[x, y]` so that only mutations with VAF in that range are actually used to determine the TIN/ TIT levels of the input.

cutoff_miscalled_clonal
An upper bound on the VAF of a cluster in the tumour data. Clusters above this value will be considered miscalled clonal clusters (e.g., due to LOH etc.).

cutoff_lv_assignment
Consider only latent variables with responsibilities above this cutoff.

N
If there are more than `N` mutations in VAF range `VAF_range_tumour`, a random subset of size `N` is retained.

FAST
If `TRUE`, it runs the analysis with reduced sampling power and accuracy. Use this to obtain a result for preliminary inspection of your data, and then run `autofit` with this parameter set to `FALSE`.

Value

An S3 object that contains the results of this analysis.

Examples

autofit(random_TIN(), FAST = TRUE)

#> Loading required package: MCMCpack

#> Loading required package: coda

#> Loading required package: MASS

#> 
#> Attaching package: ‘MASS’

#> The following object is masked from ‘package:dplyr’:
#> 
#>     select

#> ##
#> ## Markov Chain Monte Carlo Package (MCMCpack)

#> ## Copyright (C) 2003-2019 Andrew D. Martin, Kevin M. Quinn, and Jong Hee Park

#> ##
#> ## Support provided by the U.S. National Science Foundation

#> ## (Grants SES-0350646 and SES-0350613)
#> ##

#>  [ TINC ~ Load tumour/normal data ] 
#> 

#> Error in eval(lhs, parent, parent): object 'dataset' not found