Load TINC input data.

The function loads input mutations and optional copy number data for TINC. Input formats are reported at the package website.

After loading data, mutation with VAF outside a range (default [0; 0.7]) are removed from analysis. Similarly, mutations are down-sampled if in excess of some threshold (default `N=20000`).

load_TINC_input(x, cna, VAF_range_tumour = c(0, 0.7), N = 20000)

Arguments

x

A dataframe or tibble with input mutation data, reporting `chr`, `from`, `to`, `ref` and `alt`, plus `n_ref_count` and `n_alt_count`, and `t_ref_count` and `t_tot_count`.

cna

Copy Number data in the format of package CNAqc, providing `chr`, `from`, `to`, `Major` and `minor`.

VAF_range_tumour

VAF range used to filter mutations from the tumour sample.

N

If there are more than `N` mutations in VAF range `VAF_range_tumour`, a random subset of size `N` is retained.

Value

A tibble with the loaded data.

Examples

# Generating a random TIN input
rt = random_TIN()
#> ✔ Generated TINC dataset (n = 993 mutations), TIN (0.05) and TIT (1), normal and tumour coverage 30x and 120x.
#> Warning: Removed 2 rows containing missing values (`geom_bar()`).
#> Warning: Removed 2 rows containing missing values (`geom_bar()`).
load_TINC_input(x = rt$data, cna = rt$cna)
#> 
#> ── Loading TINC input data ─────────────────────────────────────────────────────
#> ✔ Input data contains n = 993 mutations, selecting operation mode.
#> ! Found CNA data, retaining only mutations that map to segments with predominant karyotype ...
#> 
#> 
#> ── CNAqc - CNA Quality Check ───────────────────────────────────────────────────
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ✔ Fortified calls for 993 somatic mutations: 993 SNVs (100%) and 0 indels.
#> ! CNAs have no CCF, assuming clonal CNAs (CCF = 1).
#> ! Added segments length (in basepairs) to CNA segments.
#> ✔ Fortified CNAs for 993 segments: 993 clonal and 0 subclonal.
#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten
#> ✔ 993 mutations mapped to clonal CNAs.
#> 
#> 
#> ── Genome coverage by karyotype, in basepairs. ──
#> 
#> # A tibble: 1 × 4
#>   minor Major     n karyotype
#>   <dbl> <dbl> <dbl> <chr>    
#> 1     1     1  2979 1:1      
#> ✔ n = 993 mutations mapped to CNA segments with karyotype 1:1 (largest available in basepairs).
#> ✔ Mutation with VAF within 0 and 0.7 ~ n = 993.
#> $mutations
#> # A tibble: 993 × 12
#>    chr        from        to ref   alt   n_ref_count n_alt_count t_ref_count
#>    <chr>     <int>     <dbl> <chr> <chr>       <dbl>       <dbl>       <dbl>
#>  1 chr6   69891882  69891883 A     T              27           0          54
#>  2 chr16  79341649  79341650 A     G              36           0          62
#>  3 chr17  65866385  65866386 T     C              28           0          72
#>  4 chr22  20239983  20239984 C     C              28           0          65
#>  5 chr8   18486211  18486212 A     C              28           0          56
#>  6 chr10  28199619  28199620 T     C              24           0          57
#>  7 chr3  109757866 109757867 G     C              17           0          55
#>  8 chr13   7117394   7117395 A     A              24           0          75
#>  9 chr3  185536924 185536925 T     T              25           0          73
#> 10 chr22  31627373  31627374 G     T              37           0          65
#> # ℹ 983 more rows
#> # ℹ 4 more variables: t_alt_count <dbl>, karyotype <chr>, id <chr>,
#> #   OK_tumour <lgl>
#> 
#> $cna_map
#> ── [ CNAqc ] MySample 993 mutations in 993 segments (993 clonal, 0 subclonal). G
#> 
#> ── Clonal CNAs 
#> 
#>  1:1  [n = 993, L =   0 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■
#> 
#> ℹ Sample Purity: 80% ~ Ploidy: 2.
#> 
#> $what_we_used
#> [1] "1:1"
#>