Load TINC input data.

The function loads input mutations and optional copy number data for TINC. Input formats are reported at the package website.

After loading data, mutation with VAF outside a range (default [0; 0.7]) are removed from analysis. Similarly, mutations are down-sampled if in excess of some threshold (default `N=20000`).

load_TINC_input(x, cna, VAF_range_tumour = c(0, 0.7), N = 20000)

Arguments

x

A dataframe or tibble with input mutation data, reporting `chr`, `from`, `to`, `ref` and `alt`, plus `n_ref_count` and `n_alt_count`, and `t_ref_count` and `t_tot_count`.

cna

Copy Number data in the format of package CNAqc, providing `chr`, `from`, `to`, `Major` and `minor`.

VAF_range_tumour

VAF range used to filter mutations from the tumour sample.

N

If there are more than `N` mutations in VAF range `VAF_range_tumour`, a random subset of size `N` is retained.

Value

A tibble with the loaded data.

Examples

# Generating a random TIN input
rt = random_TIN()
#> ✔ Generated TINC dataset (n = 995 mutations), TIN (0.05) and TIT (1), normal and tumour coverage 30x and 120x.
#> Warning: Removed 2 rows containing missing values or values outside the scale range
#> (`geom_bar()`).
#> Warning: Removed 2 rows containing missing values or values outside the scale range
#> (`geom_bar()`).
load_TINC_input(x = rt$data, cna = rt$cna)
#> 
#> ── Loading TINC input data ─────────────────────────────────────────────────────
#> ✔ Input data contains n = 995 mutations, selecting operation mode.
#> ! Found CNA data, retaining only mutations that map to segments with predominant karyotype ...
#> 
#> 
#> ── CNAqc - CNA Quality Check ───────────────────────────────────────────────────
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ✔ Fortified calls for 995 somatic mutations: 995 SNVs (100%) and 0 indels.
#> ! CNAs have no CCF, assuming clonal CNAs (CCF = 1).
#> ! Added segments length (in basepairs) to CNA segments.
#> ✔ Fortified CNAs for 995 segments: 995 clonal and 0 subclonal.
#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten
#> ✔ 995 mutations mapped to clonal CNAs.
#> 
#> 
#> ── Genome coverage by karyotype, in basepairs. ──
#> 
#> # A tibble: 1 × 4
#>   minor Major     n karyotype
#>   <dbl> <dbl> <dbl> <chr>    
#> 1     1     1  2985 1:1      
#> ✔ n = 995 mutations mapped to CNA segments with karyotype 1:1 (largest available in basepairs).
#> ✔ Mutation with VAF within 0 and 0.7 ~ n = 992.
#> $mutations
#> # A tibble: 995 × 12
#>    chr        from        to ref   alt   n_ref_count n_alt_count t_ref_count
#>    <chr>     <int>     <dbl> <chr> <chr>       <dbl>       <dbl>       <dbl>
#>  1 chr21  14303242  14303243 T     C              26           0          66
#>  2 chr2   25406363  25406364 A     C              39           1          51
#>  3 chr15   2066510   2066511 G     G              28           0          80
#>  4 chr21  40620721  40620722 G     A              26           0          48
#>  5 chr7   67303456  67303457 T     C              22           0          56
#>  6 chr2  174560877 174560878 G     T              24           0          66
#>  7 chr5  140727080 140727081 T     C              26           0          61
#>  8 chr18  58162281  58162282 A     G              24           0          53
#>  9 chr2  241243071 241243072 G     G              33           0          57
#> 10 chr19  21805886  21805887 A     G              35           0          54
#> # ℹ 985 more rows
#> # ℹ 4 more variables: t_alt_count <dbl>, karyotype <chr>, id <chr>,
#> #   OK_tumour <lgl>
#> 
#> $cna_map
#> ── [ CNAqc ] MySample 995 mutations in 995 segments (995 clonal, 0 subclonal). G
#> 
#> ── Clonal CNAs 
#> 
#>  1:1  [n = 995, L =   0 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■
#> 
#> ℹ Sample Purity: 80% ~ Ploidy: 2.
#> 
#> $what_we_used
#> [1] "1:1"
#>